requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
moviepy
Pillow
numpy
requests
sarvamai
